% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MneTaxon.R
\name{MneTaxon}
\alias{MneTaxon}
\title{Calculate the number of diagnostic zones and the element MNE for a given taxon by layer/context
Reads a zooarchaeological database and returns a .csv file with counts; elements/zones in rows and layers/contexts in columns.}
\usage{
MneTaxon(data, zones, taxon)
}
\arguments{
\item{data}{a data frame with zooarchaeological data. Should contain "layer", "taxon", "element", and "zone" columns.}

\item{zones}{a data frame with "element", "zone", and "description" columns.}

\item{taxon}{the name of the taxon (or several taxa) for which zone counts and MNEs should be calculated.}
}
\value{
MNE for each element by layer/context, as a dataframe object and a csv file; Count of diagnostic zones for the selected taxon by layer/context.
}
\description{
Calculate the number of diagnostic zones and the element MNE for a given taxon by layer/context
Reads a zooarchaeological database and returns a .csv file with counts; elements/zones in rows and layers/contexts in columns.
}
\examples{
MneTaxon(read.csv("my_data.csv"), read.csv("my_zone_list.csv"), c("capra","bovid_2"))
}
