% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NispCalc.R
\name{NispCalc}
\alias{NispCalc}
\title{Calculate NISP by taxon for each layer/context}
\usage{
NispCalc(data)
}
\arguments{
\item{data}{a data frame with zooarchaeological data. Should contain "layer", "taxon", "element", and "zone" columns.}
}
\value{
NISP per taxon by layer/context, as a csv file.
}
\description{
Reads a zooarchaeological database and returns a .csv file with taxain rows and layers/contexts in columns.
}
\examples{
NispCalc(read.csv("my_data.csv"))
}
