% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NispElement.R
\name{NispElement}
\alias{NispElement}
\title{NispElement}
\usage{
NispElement(data, taxon)
}
\arguments{
\item{data}{a data frame with zooarchaeological data. Should contain "layer", "taxon", "element", and "zone" columns.}

\item{taxon}{the name of a target taxon}
}
\value{
NISP per element for a by layer/context, as a csv file and a data frame object.
}
\description{
Calculate NISP for each element of a taxon for each layer/context. Reads a zooarchaeological database and returns a .csv file with counts of elements (in rows) by layers/contexts (in columns).
}
\examples{
NispElement(read.csv("my_data.csv", "sheep")) #add fileEncoding="UTF-8-BOM" to remove junk character added to first column name, or change it manually
}
