% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rahaf2_db.R
\docType{data}
\name{rahaf2_db}
\alias{rahaf2_db}
\title{Faunal data base from Nahal Rahaf 2 (NR2)}
\format{
An object of class \code{data.frame} with 1018 rows and 39 columns.
}
\usage{
data(rahaf2_db)
}
\description{
The site was excavated by Maayan Shemer, Nimrod Marom and Omry Barzilai in 2019-2020. The bones were identfied by Dariya Lokshin, archaeozoology Laboratory, Haifa. Part of the DEADSEA_ECO ERC Project.
}
\references{
in prep.
}
\keyword{datasets}
