% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rahaf2_zl.R
\docType{data}
\name{rahaf2_zl}
\alias{rahaf2_zl}
\title{List of diagnostic zones used in the coding of the NR2 fauna}
\format{
An object of class \code{data.frame} with 75 rows and 3 columns.
}
\usage{
data(rahaf2_zl)
}
\description{
The site was excavated by Maayan Shemer, Nimrod Marom and Omry Barzilai in 2019-2020. The bones were identfied by Dariya Lokshin, archaeozoology Laboratory, Haifa. Part of the DEADSEA_ECO ERC Project.
}
\references{
in prep.
}
\keyword{datasets}
