% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NispTaxon.R
\name{NispTaxon}
\alias{NispTaxon}
\title{NispTaxon}
\usage{
NispTaxon(data)
}
\arguments{
\item{data}{a data frame with zooarchaeological data. Should contain "layer", "taxon", "element", and "zone" columns.}
}
\value{
NISP per taxon by layer/context, as a csv file.
}
\description{
Calculate NISP by taxon for each layer/context. Reads a zooarchaeological database and returns a .csv file with taxain rows and layers/contexts in columns.
}
\examples{
NispTaxon(read.csv("my_data.csv")) #junk characters added to the first colname when file is in UTF-8; rename manually or import data with fileEncoding="UTF-8-BOM"
}
